requirements.txt
openai>=1.0
anthropic>=0.40
google-genai>=1.0
httpx>=0.27
mcp>=1.0
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
python-dotenv>=1.0
numpy>=1.24
pandas>=2.0
pyomo>=6.0
exa-py>=1.0
requests>=2.31
beautifulsoup4>=4.12
aiohttp>=3.9
loguru>=0.7
